backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alibabacloud-cas20200407==1.3.0
alibabacloud-domain20180129==3.14.13
alibabacloud-tea==0.3.6
alibabacloud_credentials==0.3.3
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.1
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.8
alibabacloud_tea_util==0.3.11
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-bssopenapi==1.7.19
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-ecs==4.24.23
amqp==5.2.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
billiard==4.2.0
boto3==1.28.80
botocore==1.31.85
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.7
daphne==3.0.2
DBUtils==3.1.0
Django==4.2.13
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-comment-migrate==0.1.7
django-cors-headers==4.3.1
django-filter==22.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-restql==0.15.3
django-simple-captcha==0.6.0
django-tenants==3.5.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==0.10.0
kombu==5.3.7
kubernetes==30.1.0
ldap3==2.9.1
MarkupSafe==2.1.5
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
paramiko==3.3.1
passlib==1.7.4
pillow==10.3.0
prompt_toolkit==3.0.45
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyPEG2==2.15.2
pypinyin==0.48.0
python-crontab==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-whois==0.9.4
pytz==2021.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.7.0
service-identity==24.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.0
supervisor==4.2.5
tornado==6.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
webssh==1.6.2
whitenoise==6.6.0
xmltodict==0.14.2
yarl==1.9.4
zope.interface==6.4.post2
